+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/CO2FootprintConfig.groovy
@@ -227,7 +227,7 @@ class CO2FootprintConfig {
     SortedMap<String, Object> collectCO2CalcOptions() {
         return [
                 "location": location,
-                "ci": ci,
+                "ci": (ci instanceof Closure) ? "dynamic" : ci,
                 "ciMarket": ciMarket,
                 "pue": pue,
                 "powerdrawMem": powerdrawMem,