+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/DataContainers/CIDataMatrix.groovy
@@ -67,18 +67,18 @@ class CIDataMatrix extends DataMatrix {
         try {
             ci = this.get(targetZone, this.ciColumn)
             log.info(Markers.unique, 
-                    "🔁 Using carbon intensity for ${HelperFunctions.bold(targetZone)} from fallback table: ${HelperFunctions.bold(ci.toString())} gCO₂eq/kWh.",
+                    "Using carbon intensity for ${HelperFunctions.bold(targetZone)} from fallback table: ${HelperFunctions.bold(ci.toString())} gCO₂eq/kWh.",
                     'using-ci-from-table-info'
                     )
         } catch (IllegalArgumentException e) {
             if (targetZone == 'GLOBAL') {
-                Exception err = new IllegalStateException("Could not retrieve ${HelperFunctions.bold('GLOBAL')} carbon intensity value from fallback table.")
+                Exception err = new IllegalStateException("Could not retrieve GLOBAL carbon intensity value from fallback table.")
                 log.error(err.getMessage(), err)
                 throw err  // <-- will stop execution
             }
             else {
                 log.warn(Markers.unique, 
-                        "🔁 Could not find carbon intensity for zone ${HelperFunctions.bold(targetZone)}: ${e.message}",
+                        "Could not find carbon intensity for zone ${targetZone.toUpperCase()}: ${e.message}",
                         'missing-ci-in-table-warning'
                 ) 
             }