+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/DataContainers/TDPDataMatrix.groovy
@@ -143,10 +143,11 @@ class TDPDataMatrix extends DataMatrix {
         }
         else if (fallbackToDefault) {
             modelData = select([this.fallbackModel] as LinkedHashSet)
+            String modelMatch = originalModel == null ? "No CPU model detected." : "Could not find CPU model \"${originalModel}\" in given TDP data table."
             log.warn(
                     Markers.unique,
-                    "Could not find CPU model \"${originalModel}\" in given TDP data table. " +
-                    "Using ${this.fallbackModel} CPU power draw value (${getTDP(modelData)} W).\n" +
+                    modelMatch +
+                    " Using ${this.fallbackModel} CPU power draw value (${getTDP(modelData)} W).\n" +
                     '\t🔖 You can help by reporting this warning and your `.nextflow.log` file to ' +
                     'https://github.com/nextflow-io/nf-co2footprint/issues/new?template=missing_chip.yaml. Thanks.'
             )