+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+# Version 1.0.0-rc.4
+## Bug Fixes:
+- Wrong memory fallback value (current machine maximum memory) now removed
+
+## Misc:
+- "🔁" now prepended for every `Markers.unique` message
+
 # Version 1.0.0-rc.3
 ## Bug Fixes:
 - Core logging no longer hidden after plugin invocation